Separate user tests from MenuTests.java
